feat: Page 내부 로직을 CardListSection 컴포넌트로 이동
